Moved smartpooling to separate layer
